requirements.txt
alabaster==0.7.16
altgraph==0.17.4
Babel==2.15.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docutils==0.21.2
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.4
lxml==5.2.2
macholib==1.16.3
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
requests==2.32.3
setuptools==70.3.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
tqdm==4.66.5
unittest-xml-reporting==3.2.0
urllib3==2.2.2
